requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
APScheduler==3.6.3
attrs==25.1.0
cachetools==4.2.2
certifi==2024.12.14
charset-normalizer==3.4.1
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
python-dotenv==1.0.1
python-telegram-bot==13.14
pytz==2024.2
requests==2.32.3
setuptools==75.8.0
six==1.17.0
tornado==6.1
tzlocal==5.2
urllib3==1.26.15
yarl==1.18.3
